pyproject.toml
[tool.poetry]
name = "ingemax-docs"
version = "0.1.0"
description = "Documentación interna de desarrollo"
authors = ["Sergio Torrado <sergio.torrado@gerpro.co>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
psycopg2 = {version = "^2.8.4", extras = ["production"]}
sphinx = "^2.4.3"
sphinx-rtd-theme = "^0.4.3"
django-redis = "^4.11.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
django-extensions = "^2.2.8"
ipython = "^7.13.0"
psycopg2-binary = "^2.8.4"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip = "migrations"
default_section = "THIRDPARTY"
known_first_party = []
known_django = "django"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
combine_as_imports = true
include_trailing_comma = true
not_skip = "__init__.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

